+++ src/controller/Movement_controller.h
@@ -20,7 +20,7 @@ public :
 
   virtual ~Movement_controller_delegate(){}
 
-  virtual void reinit(double pitch_stick, double roll_stick, double height, double paw_spreading) = 0;
+  virtual void reinit(float pitch_stick, float roll_stick, float height, float paw_spreading) = 0;
 
 };
 
@@ -35,17 +35,17 @@ public:
 	bool is_a_new_movement();
 	void run_controller();
 
-	double get_paw_spreading() const { return m_paw_spreading; }
+	float get_paw_spreading() const { return m_paw_spreading; }
 	Incline_coef_t get_incline_coef() const { return m_incline_coef; }
-	double get_center_height() const { return m_center_height; }
-	double get_pitch_stick() const { return m_incline_pitch_lin_value; }
-	double get_roll_stick() const { return m_incline_roll_lin_value; }
+	float get_center_height() const { return m_center_height; }
+	float get_pitch_stick() const { return m_incline_pitch_lin_value; }
+	float get_roll_stick() const { return m_incline_roll_lin_value; }
 
 	Movement* get_movement() { return m_movement; }
 
-	void set_new_paw_spreading(double p_paw_spreading);
-	void set_new_center_height(double p_center_height);
-	void set_new_incline(double p_pitch_stick, double p_roll_stick);
+	void set_new_paw_spreading(float p_paw_spreading);
+	void set_new_center_height(float p_center_height);
+	void set_new_incline(float p_pitch_stick, float p_roll_stick);
 	void go_back_to_default_position();
 
 	//delegate
@@ -56,15 +56,15 @@ private:
 	Movement *m_movement;
 	Movement_controller_delegate* m_delegate;
 
-	double m_paw_spreading;
-	double m_center_height;
+	float m_paw_spreading;
+	float m_center_height;
 
 	int m_current_step_number;
 	int m_step_number;
 
 	bool new_movement;
 
-	double make_more_linear(int stick_value); //out between [-1 ; 1]
+	float make_more_linear(int stick_value); //out between [-1 ; 1]
 	void make_sticks_more_linear();
 
 private:
@@ -74,10 +74,10 @@ private:
 	int m_incline_pitch_value;
 	int m_incline_roll_value;
 
-	double m_movement_x_lin_value;
-	double m_movement_y_lin_value;
-	double m_incline_pitch_lin_value;
-	double m_incline_roll_lin_value;
+	float m_movement_x_lin_value;
+	float m_movement_y_lin_value;
+	float m_incline_pitch_lin_value;
+	float m_incline_roll_lin_value;
 
 	bool m_up_pressed;
 	bool m_down_pressed;
