+++ src/movement/Movement.cpp
@@ -10,10 +10,10 @@
 #include "config.h"
 #include <cmath>
 
-Movement::Movement(Movement_type type, Movement_direction direction, double distance, double angle, int step_number) :
-	m_type(type), m_direction(direction), m_distance(distance), m_corrected_distance(0),
+Movement::Movement(Movement_type type, Movement_direction direction, float distance, float angle, int step_number) :
+	m_type(type), m_direction(direction), m_distance(distance), m_corrected_distance(0.f),
 	m_sequence_number(0), m_step_number(step_number), m_current_step_number(0),
-	m_step_distance_x(0), m_step_distance_y(0), m_step_distance_z(0),
+	m_step_distance_x(0.f), m_step_distance_y(0.f), m_step_distance_z(0.f),
 	m_angle(angle), m_paw_spreading(50), m_paw_spreading_step(NO_MOVEMENT_STEP_DIST)
 
 {
@@ -21,7 +21,7 @@ Movement::Movement(Movement_type type, Movement_direction direction, double dist
 }
 
 //try to go to a position
-double Movement::reproach_position(double present, double futur, double p_step_distance)
+float Movement::reproach_position(float present, float futur, float p_step_distance)
 {
 	if((present - futur) <= - p_step_distance)
 	{
@@ -38,15 +38,15 @@ double Movement::reproach_position(double present, double futur, double p_step_d
 }
 
 //get up and after down the paw to go to the correct height
-double Movement::get_up_paw(double final_height, Paw &paw, double p_step_distance)
+float Movement::get_up_paw(float final_height, Paw &paw, float p_step_distance)
 {
-	double z;
+	float z;
 
 	if(m_current_step_number <= (m_step_number / 2))
 	{
 		if(paw.m_current_coords.z <= MAX_HEIGHT_GET_UP)
 		{
-			z = paw.m_current_coords.z + p_step_distance*2.5;
+			z = paw.m_current_coords.z + p_step_distance*2.5f;
 		}
 		else
 			z = paw.m_current_coords.z;
@@ -54,26 +54,26 @@ double Movement::get_up_paw(double final_height, Paw &paw, double p_step_distanc
 	else
 	{
 		//compute step distance to get a linear movement
-		double step_distance_get_up = abs((final_height - paw.m_current_coords.z) / (m_step_number - m_current_step_number));
+		float step_distance_get_up = abs((final_height - paw.m_current_coords.z) / (m_step_number - m_current_step_number));
 		z = paw.m_current_coords.z - step_distance_get_up;
 	}
 
 	return z;
 }
 
-double Movement::just_get_up_paw(Paw &paw, double p_step_distance)
+float Movement::just_get_up_paw(Paw &paw, float p_step_distance)
 {
-	double z;
+	float z;
 
 	if(paw.m_current_coords.z <= MAX_HEIGHT_GET_UP)
-		z = paw.m_current_coords.z + p_step_distance*2.5;
+		z = paw.m_current_coords.z + p_step_distance*2.5f;
 	else
 		z = paw.m_current_coords.z;
 
 	return z;
 }
 
-double Movement::just_get_down_paw(double final_height, Paw &paw, double p_step_distance)
+float Movement::just_get_down_paw(float final_height, Paw &paw, float p_step_distance)
 {
 	if((paw.m_current_coords.z - final_height) <= - p_step_distance)
 	{
@@ -90,10 +90,10 @@ double Movement::just_get_down_paw(double final_height, Paw &paw, double p_step_
 }
 
 //must go to the position
-double Movement::goto_position(double present, double futur, double nb_step)
+float Movement::goto_position(float present, float futur, float nb_step)
 {
 	//compute step distance to get a linear movement
-	double step_distance = abs(present - futur) / nb_step;
+	float step_distance = abs(present - futur) / nb_step;
 
 	if(futur > present)
 	{
@@ -127,7 +127,7 @@ void Movement::compute_z_value_for_standard_paw(Side &side, Incline_coef_t p_inc
 									  p_incline_coef.C;
 }
 
-void Movement::memorize_parameters(int sequence_number, Incline_coef_t p_incline_coef, double paw_spreading)
+void Movement::memorize_parameters(int sequence_number, Incline_coef_t p_incline_coef, float paw_spreading)
 {
 	m_sequence_number = sequence_number;
 	m_incline_coef = p_incline_coef;
